Major Update
Breaking Changes
- Overhauled and Modularized the CMS Email Component
- Added Email SMTP OAuth funcionality
- Minor bugfixes throughout the application
- Moved PHPMailer to use composer
- Now using composer's autoloading

+++ kanso/framework/file/Filesystem.php
@@ -179,7 +179,12 @@ class Filesystem
 	 */
 	public static function delete(string $file): bool
 	{
-		return unlink($file);
+		if (self::exists($file) && self::isFile($file))
+		{
+			return unlink($file);
+		}
+
+		return false;
 	}
 
 	/**
